Fix symmetric evaluations for the trainer
- Fix: Return scores from the point of view of the mover.
- Use double values for the scores (avoid precission loss).

+++ src/main/java/com/joansala/engine/doe/DOE.java
@@ -338,7 +338,7 @@ public class DOE extends BaseEngine {
      * @param scorer    Evaluation function
      */
     private void evaluate(DOENode node, DOEScorer scorer) {
-        int score = scorer.apply(node.moves);
+        double score = scorer.apply(node.moves);
 
         synchronized (lock) {
             node.evaluated = true;
